+++ window/src/lib.rs
@@ -323,6 +323,10 @@ pub trait WindowOps {
     fn restore(&self) {}
     fn focus(&self) {}
 
+    fn focus_with_token(&self, _token: String) {
+        self.focus()
+    }
+
     fn toggle_fullscreen(&self) {}
 
     fn config_did_change(&self, _config: &config::ConfigHandle) {}